requirements.txt
configparser>=7.1.0
pycryptodome>=3.21.0
requests>=2.32.3